Moved config files to separate directories
I've decided that it was a stupid idea, to just use single files
and symlink them to each program's config directory.
Now I'm just gonna have directory for each program and inside is
the normal program config. And the name of the directory
is also the same the program uses, so that all those directories
can be symlinked inside .config (except for peaclock -
it's .peaclock in home directory).

Also I've removed some not needed files.